% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eRgot}
\alias{eRgot}
\title{Create spatial networks use \code{sfnetworks} for use with \code{tidygraphs}}
\usage{
eRgot(x)
}
\description{
Load, create spatial objects, clean and create spatial networks from HyPhy raw data and
}
