% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkelPy.R
\name{load_json_files}
\alias{load_json_files}
\title{Load JSON files from a folder}
\usage{
load_json_files(base_dir)
}
\arguments{
\item{base_dir}{Folder that contains .json files}
}
\value{
Named character vector of JSON file paths
}
\description{
Load JSON files from a folder
}
