% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{networkR}
\alias{networkR}
\title{Convert a spatial HyPhy object into a network, including network cleaning and simplification}
\usage{
networkR(x, organ)
}
\arguments{
\item{x}{an sf object, the output from eRgot, containing both sclerotiaPrimordia and hyphal networks}

\item{organ}{character, the organ type to process. Character, one of 'sclerotiaPrimordia' or 'hyphae'}
}
\description{
Given HyPhy data which has been converted into a spatial object via \code{HyPhy} create a spatial network,
and simplify it to remove internal nodes along un-branched edges.
}
