% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{R_get}
\alias{R_get}
\title{Convert a HyPhy skeletonized JSON object into an sf object}
\usage{
R_get(x)
}
\arguments{
\item{x}{a HyPhy object read in from jsonlite::fromJSON}
}
\description{
Converts an object in the HyPhy 'SkeletonizationTool\Skeletons\Calculations' dir
into an sf object. Allowing for downstream analysis in sf networks and facilitates
plotting with ggplot2.
}
