% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkelPy.R
\name{run_fyskel_pipeline}
\alias{run_fyskel_pipeline}
\title{Run full FySkel pipeline on a folder of JSON files}
\usage{
run_fyskel_pipeline(base_dir = NULL, output_dir = getwd())
}
\arguments{
\item{base_dir}{Path to folder with JSON files (optional)}

\item{output_dir}{Output directory for CSV + PNG files}
}
\value{
A list with results and paths
}
\description{
Run full FySkel pipeline on a folder of JSON files
}
