% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkelPy.R
\name{safe_eRgot}
\alias{safe_eRgot}
\title{Safe wrapper around eRgot(), prevents crashes from early timepoints with no networks}
\usage{
safe_eRgot(path)
}
\description{
Safe wrapper around eRgot(), prevents crashes from early timepoints with no networks
}
