% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkelPy.R
\name{select_json_folder}
\alias{select_json_folder}
\title{Select folder containing JSON files}
\usage{
select_json_folder()
}
\value{
A character string: path to folder
}
\description{
Uses RStudio dialog if available, otherwise terminal input.
}
