% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unclean_mapR}
\alias{unclean_mapR}
\title{Plot an imported HyPhy object}
\usage{
unclean_mapR(x)
}
\arguments{
\item{x}{an sf object, the output from eRgot, containing both sclerotiaPrimordia and hyphal networks}
}
\description{
Plots HyPhy imported data
}
